Tearing out main activity page
