+++ app/src/main/java/com/example/hushed/HushedApplication.java
@@ -26,7 +26,7 @@ public class HushedApplication
     @Override public void onActivityPaused(@NonNull Activity activity) { }
     @Override public void onActivityStopped(@NonNull Activity activity) {
         Log.i("Test", "Activity " + activity.getClass() + " Stopped.");
-        if (activity instanceof MainActivity) {
+        if (activity instanceof MessageActivity) {
             DataSource.Companion.saveTo(getSharedPreferences("DataSource", Context.MODE_PRIVATE));
         }
     }
